requirements.txt
alembic==1.4.2
cffi==1.14.0
chardet==2.3.0
click==7.1.2
cryptography==1.7.1
dnspython==1.16.0
email-validator==1.1.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
gunicorn==20.0.4
httplib2==0.9.2
idna==2.2
itsdangerous==1.1.0
Jinja2==2.11.2
keyring==10.1
keyrings.alt==1.3
Mako==1.1.2
MarkupSafe==1.1.1
numpy==1.18.4
Pillow==7.1.2
pyasn1==0.1.9
pycparser==2.20
PyMySQL==0.9.3
python-dateutil==2.8.1
python-debian==0.1.30
python-dotenv==0.13.0
python-editor==1.0.4
pywin32-ctypes==0.2.0
pyxdg==0.25
requests==2.12.4
SecretStorage==2.3.1
six==1.10.0
SQLAlchemy==1.3.17
SQLAlchemy-ImageAttach==1.1.0
urllib3==1.19.1
Wand==0.6.1
Werkzeug==1.0.1
WTForms==2.3.1
